+++ aeron-driver/src/main/java/uk/co/real_logic/aeron/driver/ClientProxy.java
@@ -69,8 +69,8 @@ public class ClientProxy
     }
 
     public void onNewTermBuffers(final int msgTypeId,
-                                 final long sessionId,
-                                 final long channelId,
+                                 final int sessionId,
+                                 final int channelId,
                                  final int termId,
                                  final String destination,
                                  final TermBuffers termBuffers,
